Adds upload and retrieval of rules
Instead of using a static file, this change allows the rules to be
uploaded to the /rules endpoint as the body of a post, and retrieved as
a get.  This uses the environment variable `SHOUT_ADMIN_AUTH` to specify
the basic auth, which defaults to "shout:shout".

This removed the ability to specify rules by providing a file.  Instead
you can use curl to directly upload the rules from the file.