Rules Upload Endpoint (POST /rules)
Instead of using a static file, Shout! now expects notification rules to be
uploaded to the /rules endpoint as the body of a POST, and retrieved as
a GET, subject to HTTP Basic Authentication.  Administrator credentials
can be supplied via the `$SHOUT_ADMIN_AUTH` environment variable,
in the format "username:password".  Defaults to "shout:shout".

Rules can no longer be loaded from a file.